# perf: 50% faster aqi compute on single var