## Single var

```bash
go test -benchmem -run=^$ -bench ^BenchmarkAlgoCalcOnSingleDataset$ github.com/ringsaturn/go-aqi/mep ./... > old.txt
```

```
goos: darwin
goarch: arm64
pkg: github.com/ringsaturn/go-aqi/mep
BenchmarkAlgoCalcOnSingleDataset-16    	12619250	        94.83 ns/op	      56 B/op	       3 allocs/op
PASS
ok  	github.com/ringsaturn/go-aqi/mep	2.531s
PASS
ok  	github.com/ringsaturn/go-aqi	0.424s
PASS
ok  	github.com/ringsaturn/go-aqi/epa	0.428s
```

```bash
go test -benchmem -run=^$ -bench ^BenchmarkAlgoCalcOnSingleDataset$ github.com/ringsaturn/go-aqi/mep ./... > new.txt
```

```
goos: darwin
goarch: arm64
pkg: github.com/ringsaturn/go-aqi/mep
BenchmarkAlgoCalcOnSingleDataset-16    	26883134	        44.82 ns/op	      28 B/op	       3 allocs/op
PASS
ok  	github.com/ringsaturn/go-aqi/mep	2.666s
PASS
ok  	github.com/ringsaturn/go-aqi	0.498s
PASS
ok  	github.com/ringsaturn/go-aqi/epa	0.425s
```

```
➜  go-aqi git:(ringsaturn/perf) ✗ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/ringsaturn/go-aqi/mep
                           │   old.txt    │             new.txt             │
                           │    sec/op    │    sec/op     vs base           │
AlgoCalcOnSingleDataset-16   94.83n ± ∞ ¹   44.82n ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                           │   old.txt   │            new.txt             │
                           │    B/op     │    B/op      vs base           │
AlgoCalcOnSingleDataset-16   56.00 ± ∞ ¹   28.00 ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                           │   old.txt   │            new.txt             │
                           │  allocs/op  │  allocs/op   vs base           │
AlgoCalcOnSingleDataset-16   3.000 ± ∞ ¹   3.000 ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
```

## Multi Var 384 steps:

```
benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/ringsaturn/go-aqi/mep
                       │   old.txt    │             new.txt             │
                       │    sec/op    │    sec/op     vs base           │
AlgoCalcOnLargeData-16   82.76µ ± ∞ ¹   68.59µ ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                       │    old.txt    │             new.txt              │
                       │     B/op      │     B/op       vs base           │
AlgoCalcOnLargeData-16   21.00Ki ± ∞ ¹   63.00Ki ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                       │   old.txt    │             new.txt             │
                       │  allocs/op   │  allocs/op    vs base           │
AlgoCalcOnLargeData-16   1.152k ± ∞ ¹   3.072k ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

`old.txt`:

```
goos: darwin
goarch: arm64
pkg: github.com/ringsaturn/go-aqi/mep
BenchmarkAlgoCalcOnLargeData-16    	   14083	     82763 ns/op	   21504 B/op	    1152 allocs/op
PASS
ok  	github.com/ringsaturn/go-aqi/mep	2.245s
PASS
ok  	github.com/ringsaturn/go-aqi	0.302s
PASS
ok  	github.com/ringsaturn/go-aqi/epa	0.328s
```


`new.txt`:

```
goos: darwin
goarch: arm64
pkg: github.com/ringsaturn/go-aqi/mep
BenchmarkAlgoCalcOnLargeData-16    	   18598	     68586 ns/op	   64512 B/op	    3072 allocs/op
PASS
ok  	github.com/ringsaturn/go-aqi/mep	2.413s
PASS
ok  	github.com/ringsaturn/go-aqi	0.403s
PASS
ok  	github.com/ringsaturn/go-aqi/epa	0.415s
```